2022/06/rundeck-recipe-repository/pyproject.toml
[tool.poetry]
name = "rundeck-recipe-repository"
version = "0.1.0"
description = ""
authors = ["Willian Antunes <willian.lima.antunes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyGithub = "^1.55"
django = "*"
python-sonarqube-api = "*"
### Tooling
requests = "^2"

[tool.poetry.dev-dependencies]
### Pytest: Essentials
pytest-mock = "^3"
pytest-django = "^4"
pytest-cov = "^3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target_version = ["py310"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | migrations
    | settings.py
  )/
)
'''
